Added sinatra reloader

+++ app.rb
@@ -14,23 +14,26 @@ end
 
 #:id - Dynamic URL attribute
 get '/todos/:id' do
+  # "SHOW"
   #puts params #Full hash
   #puts params[:id] #Id only
-  # "SHOW"
-  id  = params[:id].to_f
+  id  = params[:id].to_i
   TodoManager.show(id)
 end
 
 put '/todos/:id' do
   # "UPDATE"
-  id = params[:id].to_f
+  id = params[:id].to_i
   to_add = params[:todo]
   TodoManager.update(id, to_add)
 end
+# "?todo=task&" - Query Parameters, separted by and for multiple params
+
+
 
 delete '/todos/:id' do
   # "DESTROY"
-  id = params[:id].to_f
+  id = params[:id].to_i
   TodoManager.delete(id)
 end
